requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==10.0.0
pytesseract==0.3.10
requests==2.31.0
Werkzeug==2.3.7